backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
mysql-connector-python==8.1.0
redis==4.6.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==6.0.0
groq==0.29.0
openai==1.3.7
sentry-sdk[flask]==1.38.0
prometheus-client==0.17.1
gunicorn==22.0.0
flask-sock==0.7.0
gevent==25.5.1
selenium==4.33.0
langchain>=0.3.0
langchain_community
readability-lxml
playwright
langchain_groq
